refactor: split audio module
