+++ src/audio/player.rs
@@ -1,12 +1,13 @@
-use std::{
-    io::Cursor,
-    sync::Arc,
-    time::{Duration, Instant},
-};
+use std::sync::Arc;
+use std::time::{Duration, Instant};
 
-pub mod output;
+use rodio::{OutputStream, Sink, Source};
 
-use rodio::{Decoder, OutputStream, Sink, Source};
+use super::DEFAULT_ANTI_CLIP_FADE;
+use super::fade::{EdgeFade, FadeOutHandle, FadeOutOnRequest, fade_duration, fade_frames_for_duration};
+use super::mixer::{decoder_duration, decoder_from_bytes, map_seek_error, wav_header_duration};
+use super::output::{AudioOutputConfig, ResolvedOutput, open_output_stream};
+use super::routing::{duration_from_secs_f32, duration_mod};
 
 /// Simple audio helper that plays a loaded wav buffer and reports progress.
 pub struct AudioPlayer {
@@ -27,8 +28,6 @@ pub struct AudioPlayer {
     #[cfg(test)]
     elapsed_override: Option<Duration>,
 }
-
-const DEFAULT_ANTI_CLIP_FADE: Duration = Duration::from_millis(2);
 
 impl AudioPlayer {
     /// Create a new audio player using the default output device.
@@ -63,8 +62,8 @@ impl AudioPlayer {
     pub fn set_audio(&mut self, data: Vec<u8>, duration: f32) {
         let audio = Arc::from(data);
         let provided = duration.max(0.0);
-        let fallback = Self::decoder_duration(&audio)
-            .or_else(|| Self::wav_header_duration(&audio))
+        let fallback = decoder_duration(&audio)
+            .or_else(|| wav_header_duration(&audio))
             .unwrap_or(0.0);
         let chosen = if provided > 0.0 { provided } else { fallback };
         self.track_duration = Some(chosen);
@@ -152,7 +151,7 @@ impl AudioPlayer {
         self.fade_out_current_sink(self.anti_clip_fade());
 
         let fade = fade_duration(duration, self.anti_clip_fade());
-        let source = Self::decoder_from_bytes(bytes)?;
+        let source = decoder_from_bytes(bytes)?;
         let limited = source
             .fade_in(fade)
             .take_duration(Duration::from_secs_f32(duration))
@@ -271,10 +270,10 @@ impl AudioPlayer {
 
         self.fade_out_current_sink(self.anti_clip_fade());
 
-        let mut source = Self::decoder_from_bytes(bytes)?;
+        let mut source = decoder_from_bytes(bytes)?;
         source
             .try_seek(Duration::from_secs_f32(bounded_start))
-            .map_err(Self::map_seek_error)?;
+            .map_err(map_seek_error)?;
         let limited = source
             .fade_in(fade)
             .take_duration(Duration::from_secs_f32(span_length))
@@ -306,15 +305,6 @@ impl AudioPlayer {
         Ok(())
     }
 
-    fn map_seek_error(error: rodio::source::SeekError) -> String {
-        match error {
-            rodio::source::SeekError::NotSupported { .. } => {
-                "Seeking not supported for this audio source".into()
-            }
-            _ => format!("Audio seek failed: {error}"),
-        }
-    }
-
     fn audio_bytes(&self) -> Result<Arc<[u8]>, String> {
         self.current_audio
             .as_ref()
@@ -330,47 +320,18 @@ impl AudioPlayer {
         }
     }
 
-    fn decoder_from_bytes(bytes: Arc<[u8]>) -> Result<Decoder<Cursor<Arc<[u8]>>>, String> {
-        let byte_len = bytes.len() as u64;
-        Decoder::builder()
-            .with_data(Cursor::new(bytes))
-            .with_byte_len(byte_len)
-            .with_seekable(true)
-            .with_hint("wav")
-            .build()
-            .map_err(|error| format!("Audio decode failed: {error}"))
-    }
-
-    fn decoder_duration(bytes: &Arc<[u8]>) -> Option<f32> {
-        Self::decoder_from_bytes(bytes.clone())
-            .ok()
-            .and_then(|decoder| decoder.total_duration())
-            .map(|duration| duration.as_secs_f32())
-    }
-
-    fn wav_header_duration(bytes: &Arc<[u8]>) -> Option<f32> {
-        let reader = hound::WavReader::new(Cursor::new(bytes.clone())).ok()?;
-        let spec = reader.spec();
-        let sample_rate = spec.sample_rate as f32;
-        let channels = spec.channels.max(1) as f32;
-        if sample_rate <= 0.0 {
-            return None;
-        }
-        Some(reader.duration() as f32 / (sample_rate * channels))
-    }
-
     #[cfg(test)]
-    fn span_sample_count(
+    pub(crate) fn span_sample_count(
         bytes: Arc<[u8]>,
         start_seconds: f32,
         end_seconds: f32,
     ) -> Result<(usize, u32, u16), String> {
-        let mut source = Self::decoder_from_bytes(bytes)?;
+        let mut source = decoder_from_bytes(bytes)?;
         source
             .try_seek(Duration::from_secs_f32(start_seconds))
-            .map_err(Self::map_seek_error)?;
+            .map_err(map_seek_error)?;
         let span_length = (end_seconds - start_seconds).max(0.001);
-        let fade = fade_duration(span_length, self.anti_clip_fade());
+        let fade = fade_duration(span_length, DEFAULT_ANTI_CLIP_FADE);
         let limited = source
             .fade_in(fade)
             .take_duration(Duration::from_secs_f32(span_length))
@@ -418,7 +379,7 @@ impl AudioPlayer {
     #[cfg(test)]
     /// Build a looped playing instance for tests that need an active sink.
     pub fn playing_for_tests() -> Option<Self> {
-        use rodio::source::{SineWave, Source};
+        use rodio::source::SineWave;
 
         let outcome = open_output_stream(&AudioOutputConfig::default()).ok()?;
         let sink = rodio::Sink::connect_new(outcome.stream.mixer());
@@ -447,312 +408,3 @@ impl AudioPlayer {
         })
     }
 }
-
-fn fade_duration(span_seconds: f32, max_fade: Duration) -> Duration {
-    if span_seconds <= 0.0 || max_fade.is_zero() {
-        return Duration::ZERO;
-    }
-    let max_fade_secs = max_fade.as_secs_f32();
-    let clamped = max_fade_secs.min(span_seconds * 0.5);
-    duration_from_secs_f32(clamped)
-}
-
-fn fade_frames_for_duration(sample_rate: u32, fade: Duration) -> u32 {
-    if fade.is_zero() || sample_rate == 0 {
-        return 1;
-    }
-    let frames = (fade.as_secs_f64() * sample_rate as f64).ceil();
-    frames.clamp(1.0, u32::MAX as f64) as u32
-}
-
-#[cfg(test)]
-fn normalized_progress(
-    span: Option<(f32, f32)>,
-    duration: f32,
-    elapsed: f32,
-    looping: bool,
-) -> Option<f32> {
-    if duration <= 0.0 {
-        return None;
-    }
-    let (start, end) = span.unwrap_or((0.0, duration));
-    let span_length = (end - start).max(f32::EPSILON);
-    let within_span = if looping {
-        elapsed % span_length
-    } else {
-        elapsed.min(span_length)
-    };
-    let absolute = start + within_span;
-    Some((absolute / duration).clamp(0.0, 1.0))
-}
-
-fn duration_from_secs_f32(seconds: f32) -> Duration {
-    if !seconds.is_finite() || seconds <= 0.0 {
-        return Duration::ZERO;
-    }
-    Duration::from_secs_f64(seconds as f64)
-}
-
-fn duration_mod(value: Duration, modulus: Duration) -> Duration {
-    let modulus_nanos = modulus.as_nanos();
-    if modulus_nanos == 0 {
-        return Duration::ZERO;
-    }
-    let remainder = value.as_nanos() % modulus_nanos;
-    let secs = (remainder / 1_000_000_000) as u64;
-    let nanos = (remainder % 1_000_000_000) as u32;
-    Duration::new(secs, nanos)
-}
-
-#[derive(Clone)]
-struct EdgeFade<S> {
-    inner: S,
-    fade_secs: f32,
-    total_secs: Option<f32>,
-    fade_out_start: Option<f32>,
-    sample_rate: u32,
-    channels: u16,
-    samples_emitted: u64,
-}
-
-impl<S> EdgeFade<S> {
-    fn new(inner: S, fade: Duration) -> Self
-    where
-        S: Source<Item = f32> + Clone,
-    {
-        let fade_secs = fade.as_secs_f32();
-        let total_secs = inner.total_duration().map(|d| d.as_secs_f32());
-        let fade_out_start = total_secs.and_then(|total| {
-            if fade_secs <= 0.0 || fade_secs >= total {
-                None
-            } else {
-                Some(total - fade_secs)
-            }
-        });
-        let sample_rate = inner.sample_rate();
-        let channels = inner.channels();
-        Self {
-            inner,
-            fade_secs,
-            total_secs,
-            fade_out_start,
-            sample_rate,
-            channels,
-            samples_emitted: 0,
-        }
-    }
-}
-
-impl<S> Iterator for EdgeFade<S>
-where
-    S: Source<Item = f32>,
-{
-    type Item = f32;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let sample = self.inner.next()?;
-        let time = if self.sample_rate > 0 && self.channels > 0 {
-            self.samples_emitted as f32 / (self.sample_rate as f32 * self.channels as f32)
-        } else {
-            0.0
-        };
-        self.samples_emitted = self.samples_emitted.saturating_add(1);
-        if self.fade_secs <= 0.0 {
-            return Some(sample);
-        }
-        let mut factor = 1.0;
-        if time < self.fade_secs {
-            factor *= (time / self.fade_secs).clamp(0.0, 1.0);
-        }
-        if let (Some(total), Some(start)) = (self.total_secs, self.fade_out_start)
-            && time > start
-        {
-            factor *= ((total - time) / self.fade_secs).clamp(0.0, 1.0);
-        }
-        Some(sample * factor)
-    }
-}
-
-impl<S> Source for EdgeFade<S>
-where
-    S: Source<Item = f32>,
-{
-    #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        self.inner.current_span_len()
-    }
-
-    #[inline]
-    fn channels(&self) -> u16 {
-        self.inner.channels()
-    }
-
-    #[inline]
-    fn sample_rate(&self) -> u32 {
-        self.inner.sample_rate()
-    }
-
-    #[inline]
-    fn total_duration(&self) -> Option<Duration> {
-        self.inner.total_duration()
-    }
-}
-
-#[derive(Clone)]
-struct FadeOutHandle {
-    requested_frames: Arc<std::sync::atomic::AtomicU32>,
-}
-
-impl FadeOutHandle {
-    fn new() -> Self {
-        Self {
-            requested_frames: Arc::new(std::sync::atomic::AtomicU32::new(0)),
-        }
-    }
-
-    fn request_fade_out_frames(&self, frames: u32) {
-        self.requested_frames
-            .store(frames.max(1), std::sync::atomic::Ordering::Release);
-    }
-
-    fn take_requested_frames(&self) -> Option<u32> {
-        let frames = self
-            .requested_frames
-            .swap(0, std::sync::atomic::Ordering::AcqRel);
-        if frames == 0 { None } else { Some(frames) }
-    }
-}
-
-struct FadeOutOnRequest<S> {
-    inner: S,
-    handle: FadeOutHandle,
-    sample_rate: u32,
-    channels: u16,
-    samples_emitted: u64,
-    pending_frames: Option<u32>,
-    fade_start_frame: u64,
-    fade_total_frames: u32,
-    fading: bool,
-}
-
-impl<S> FadeOutOnRequest<S>
-where
-    S: Source<Item = f32>,
-{
-    fn new(inner: S, handle: FadeOutHandle) -> Self {
-        let sample_rate = inner.sample_rate();
-        let channels = inner.channels();
-        Self {
-            inner,
-            handle,
-            sample_rate,
-            channels,
-            samples_emitted: 0,
-            pending_frames: None,
-            fade_start_frame: 0,
-            fade_total_frames: 0,
-            fading: false,
-        }
-    }
-
-    fn current_frame(&self) -> u64 {
-        let channels = self.channels.max(1) as u64;
-        self.samples_emitted / channels
-    }
-
-    fn is_frame_boundary(&self) -> bool {
-        let channels = self.channels.max(1) as u64;
-        self.samples_emitted % channels == 0
-    }
-
-    fn start_fade_if_ready(&mut self) {
-        if self.fading {
-            return;
-        }
-        if self.pending_frames.is_none() {
-            self.pending_frames = self.handle.take_requested_frames();
-        }
-        if self.pending_frames.is_none() || !self.is_frame_boundary() {
-            return;
-        }
-        let Some(frames) = self.pending_frames.take() else {
-            return;
-        };
-        self.fading = true;
-        self.fade_start_frame = self.current_frame();
-        self.fade_total_frames = frames.max(1);
-    }
-
-    fn fade_factor(&self) -> f32 {
-        if !self.fading {
-            return 1.0;
-        }
-        let current_frame = self.current_frame();
-        let offset = current_frame.saturating_sub(self.fade_start_frame);
-        let total = self.fade_total_frames.max(1) as u64;
-        if offset >= total {
-            return 0.0;
-        }
-        if total <= 1 {
-            return 0.0;
-        }
-        let denom = (total - 1) as f32;
-        let progress = (offset as f32 / denom).clamp(0.0, 1.0);
-        (1.0 - progress).clamp(0.0, 1.0)
-    }
-}
-
-impl<S> Iterator for FadeOutOnRequest<S>
-where
-    S: Source<Item = f32>,
-{
-    type Item = f32;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.start_fade_if_ready();
-        if self.fading && self.is_frame_boundary() {
-            let current_frame = self.current_frame();
-            let offset = current_frame.saturating_sub(self.fade_start_frame);
-            if offset >= self.fade_total_frames as u64 {
-                return None;
-            }
-        }
-
-        let sample = self.inner.next()?;
-        let factor = self.fade_factor();
-        self.samples_emitted = self.samples_emitted.saturating_add(1);
-        Some(sample * factor)
-    }
-}
-
-impl<S> Source for FadeOutOnRequest<S>
-where
-    S: Source<Item = f32>,
-{
-    #[inline]
-    fn current_span_len(&self) -> Option<usize> {
-        self.inner.current_span_len()
-    }
-
-    #[inline]
-    fn channels(&self) -> u16 {
-        self.channels
-    }
-
-    #[inline]
-    fn sample_rate(&self) -> u32 {
-        self.sample_rate
-    }
-
-    #[inline]
-    fn total_duration(&self) -> Option<Duration> {
-        None
-    }
-}
-
-#[cfg(test)]
-mod tests;
-pub use output::{
-    AudioDeviceSummary, AudioHostSummary, AudioOutputConfig, AudioOutputError, ResolvedOutput,
-    available_devices, available_hosts, open_output_stream, supported_sample_rates,
-};
